# Conflicts:
#	composeApp/build.gradle.kts
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/Database.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/service/modern/PlayerServiceModern.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/components/themed/Dialog.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/items/AlbumItem.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/artist/ArtistOverviewModern.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/artist/ArtistScreenModern.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/home/HomeAlbums.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/home/HomeArtists.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/home/HomeLibrary.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/playlist/PlaylistSongList.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/podcast/Podcast.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/settings/GeneralSettings.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/screens/statistics/StatisticsPage.kt
#	composeApp/src/androidMain/kotlin/it/fast4x/rimusic/utils/LocalPlaylistSongs.kt
#	composeApp/src/androidUpstream/kotlin/it/fast4x/rimusic/ui/screens/album/AlbumDetails.kt
#	composeApp/src/androidUpstream/kotlin/it/fast4x/rimusic/ui/screens/localplaylist/LocalPlaylistSongs.kt
#	composeApp/src/androidUpstream/kotlin/it/fast4x/rimusic/ui/screens/player/Queue.kt
#	updatedVersion/updatedVersion.ver
#	updatedVersion/updatedVersionCode.ver

+++ composeApp/src/androidMain/kotlin/it/fast4x/rimusic/ui/components/themed/Dialog.kt
@@ -53,7 +53,6 @@ import androidx.compose.material3.ripple
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
@@ -100,10 +99,10 @@ import it.fast4x.compose.persist.persist
 import it.fast4x.innertube.Innertube
 import it.fast4x.innertube.YtMusic
 import it.fast4x.innertube.models.bodies.SearchBody
-import it.fast4x.innertube.requests.ArtistPage
 import it.fast4x.innertube.requests.searchPage
 import it.fast4x.innertube.utils.from
 import it.fast4x.rimusic.Database
+import it.fast4x.rimusic.Database.Companion.update
 import it.fast4x.rimusic.LocalPlayerServiceBinder
 import it.fast4x.rimusic.R
 import it.fast4x.rimusic.cleanPrefix
@@ -114,13 +113,15 @@ import it.fast4x.rimusic.enums.ValidationType
 import it.fast4x.rimusic.models.Album
 import it.fast4x.rimusic.models.Artist
 import it.fast4x.rimusic.models.Info
+import it.fast4x.rimusic.models.Playlist
 import it.fast4x.rimusic.models.Song
 import it.fast4x.rimusic.models.SongAlbumMap
 import it.fast4x.rimusic.models.SongArtistMap
 import it.fast4x.rimusic.models.SongPlaylistMap
-import it.fast4x.rimusic.ui.styling.favoritesIcon
 import it.fast4x.rimusic.typography
+import it.fast4x.rimusic.ui.screens.settings.isYouTubeSyncEnabled
 import it.fast4x.rimusic.ui.styling.Dimensions
+import it.fast4x.rimusic.ui.styling.favoritesIcon
 import it.fast4x.rimusic.ui.styling.px
 import it.fast4x.rimusic.ui.styling.shimmer
 import it.fast4x.rimusic.utils.VinylSizeKey
@@ -148,6 +149,7 @@ import it.fast4x.rimusic.utils.playbackPitchKey
 import it.fast4x.rimusic.utils.playbackSpeedKey
 import it.fast4x.rimusic.utils.playbackVolumeKey
 import it.fast4x.rimusic.utils.rememberPreference
+import it.fast4x.rimusic.utils.removeYTSongFromPlaylist
 import it.fast4x.rimusic.utils.resize
 import it.fast4x.rimusic.utils.secondary
 import it.fast4x.rimusic.utils.semiBold
@@ -160,8 +162,11 @@ import it.fast4x.rimusic.utils.thumbnailFadeKey
 import it.fast4x.rimusic.utils.thumbnailRoundnessKey
 import it.fast4x.rimusic.utils.thumbnailSpacingKey
 import it.fast4x.rimusic.utils.thumbnailSpacingLKey
-import kotlinx.coroutines.delay
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.firstOrNull
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.withContext
 
@@ -554,7 +559,6 @@ inline fun SelectorArtistsDialog(
                     HorizontalPager(state = pagerState) { idArtist ->
                         val browseId = values[idArtist].id
                         var artist by persist<Artist?>("artist/$browseId/artist")
-                        var artistPage by persist<ArtistPage?>("artist/$browseId/artistPage")
                         LaunchedEffect(browseId) {
                             Database.artist(values[idArtist].id).collect{artist = it}
                         }
@@ -563,16 +567,9 @@ inline fun SelectorArtistsDialog(
                                 withContext(Dispatchers.IO) {
                                     YtMusic.getArtistPage(browseId = browseId)
                                         .onSuccess { currentArtistPage ->
-                                            artistPage = currentArtistPage
-                                            Database.upsert(
-                                                Artist(
-                                                    id = browseId,
-                                                    name = currentArtistPage.artist.info?.name,
-                                                    thumbnailUrl = currentArtistPage.artist.thumbnail?.url,
-                                                    timestamp = artist?.timestamp,
-                                                    bookmarkedAt = artist?.bookmarkedAt
-                                                )
-                                            )
+                                            artist?.copy(
+                                                thumbnailUrl = currentArtistPage.artist.thumbnail?.url
+                                            )?.let(::update)
                                             Database.artist(values[idArtist].id).collect{artist = it}
                                         }
                                 }
@@ -598,7 +595,7 @@ inline fun SelectorArtistsDialog(
                             )
                             values[idArtist].name?.let { it1 ->
                                 BasicText(
-                                    text = it1,
+                                    text = cleanPrefix(it1),
                                     maxLines = 3,
                                     overflow = TextOverflow.Ellipsis,
                                     style = typography().xs.medium,
@@ -607,7 +604,7 @@ inline fun SelectorArtistsDialog(
                                         .align(Alignment.BottomCenter)
                                 )
                                 BasicText(
-                                    text = it1,
+                                    text = cleanPrefix(it1),
                                     style = typography().xs.medium.merge(TextStyle(
                                         drawStyle = Stroke(width = 1.0f, join = StrokeJoin.Round),
                                         color = if (colorPaletteMode == ColorPaletteMode.Light || (colorPaletteMode == ColorPaletteMode.System && (!isSystemInDarkTheme()))) Color.White.copy(0.5f)
@@ -1832,6 +1829,7 @@ fun SongMatchingDialog(
     songToRematch : Song,
     playlistId : Long,
     position : Int,
+    playlist : Playlist?,
     onDismiss: (() -> Unit)
 ) {
     Dialog(
@@ -2001,6 +1999,7 @@ fun SongMatchingDialog(
                     itemsIndexed(songsList) { _, song ->
                         val artistsNames = song?.authors?.filter { it.endpoint != null }?.map { it.name }
                         val artistsIds = song?.authors?.filter { it.endpoint != null }?.map { it.endpoint?.browseId }
+                        val artistNameString = song?.asMediaItem?.mediaMetadata?.artist?.toString() ?: ""
                         if (song != null) {
                             Row(horizontalArrangement = Arrangement.Start,
                                 verticalAlignment = Alignment.CenterVertically,
@@ -2010,21 +2009,38 @@ fun SongMatchingDialog(
                                     .padding(vertical = 10.dp)
                                     .clickable(onClick = {
                                         Database.asyncTransaction {
-                                            deleteSongFromPlaylist(songToRematch.id, playlistId)
+                                            if (isYouTubeSyncEnabled() && playlist?.isYoutubePlaylist == true && playlist.isEditable){
+                                                CoroutineScope(Dispatchers.IO).launch {
+                                                    if (removeYTSongFromPlaylist(songToRematch.id, playlist.browseId ?: "", playlistId))
+                                                        deleteSongFromPlaylist(songToRematch.id, playlistId)
+                                                }
+                                            } else {
+                                                deleteSongFromPlaylist(songToRematch.id, playlistId)
+                                            }
+
                                             if (songExist(song.asSong.id) == 0) {
                                                 Database.insert(song.asMediaItem)
                                             }
+
                                             insert(
                                                 SongPlaylistMap(
                                                     songId = song.asMediaItem.mediaId,
                                                     playlistId = playlistId,
                                                     position = position
-                                                )
+                                                ).default()
                                             )
                                             insert(
                                                 Album(id = song.album?.endpoint?.browseId ?: "", title = song.asMediaItem.mediaMetadata.albumTitle?.toString()),
                                                 SongAlbumMap(songId = song.asMediaItem.mediaId, albumId = song.album?.endpoint?.browseId ?: "", position = null)
                                             )
+                                            CoroutineScope(Dispatchers.IO).launch {
+                                                val album = Database.album(song.album?.endpoint?.browseId ?: "").firstOrNull()
+                                                album?.copy(thumbnailUrl = song.thumbnail?.url)?.let { update(it) }
+
+                                                if (isYouTubeSyncEnabled() && playlist?.isYoutubePlaylist == true && playlist.isEditable){
+                                                    YtMusic.addToPlaylist(playlist.browseId ?: "", song.asMediaItem.mediaId)
+                                                }
+                                            }
                                             if ((artistsNames != null) && (artistsIds != null)) {
                                                 artistsNames.let { artistNames ->
                                                     artistsIds.let { artistIds ->
@@ -2044,6 +2060,7 @@ fun SongMatchingDialog(
                                                     }
                                                 }
                                             }
+                                            Database.updateSongArtist(song.asMediaItem.mediaId, artistNameString)
                                         }
                                         onDismiss()
                                     }
